# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "WebApiEngine",
    "version": "1.0"
  },
  "paths": {
...
    "/Form/{formId}/draft": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedRevision",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Form/{formId}/publish": {
      "post": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/Form/{formId}/{formIdentifier}": {
      "get": {
        "tags": [
          "Form"
        ],
        "parameters": [
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "formIdentifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormDtoApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/form-section": {
      "get": {
        "tags": [
          "FormSection"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoArrayApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoArrayApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoArrayApiStatusResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FormSection"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormSectionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormSectionRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/CreateFormSectionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              }
            }
          }
        }
      }
    },
    "/form-section/{sectionId}": {
      "get": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFormSectionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFormSectionRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/RenameFormSectionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionMetadataDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/form-section/{sectionId}/versions": {
      "get": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionSummaryDtoArrayApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionSummaryDtoArrayApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionSummaryDtoArrayApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/form-section/{sectionId}/versions/{version}": {
      "get": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/form-section/{sectionId}/draft": {
      "get": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormSectionAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormSectionAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SaveFormSectionAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionAuthoringDraftDtoApiStatusResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expectedRevision",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatusResult"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/form-section/{sectionId}/publish": {
      "post": {
        "tags": [
          "FormSection"
        ],
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormSectionAuthoringDraftRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormSectionAuthoringDraftRequestDto"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/PublishFormSectionAuthoringDraftRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormSectionVersionDtoApiStatusResult"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiValidationProblem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ApiProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "successful": {
            "type": "boolean",
            "readOnly": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {}
      },
      "ApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApiValidationProblem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "successful": {
            "type": "boolean",
            "readOnly": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {}
      },
...
      "CreateFormSectionRequestDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FormAuthoringDraftDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormAuthoringDraftDto"
          }
        },
        "additionalProperties": false
      },
...
      "FormDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormDto"
          }
        },
        "additionalProperties": false
      },
...
      "FormSectionAuthoringDraftDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormSectionAuthoringDraftDto"
          }
        },
        "additionalProperties": false
      },
...
      "FormSectionMetadataDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormSectionMetadataDto"
          }
        },
        "additionalProperties": false
      },
      "FormSectionMetadataDtoArrayApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormSectionMetadataDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FormSectionVersionDtoApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "$ref": "#/components/schemas/FormSectionVersionDto"
          }
        },
        "additionalProperties": false
      },
...
      "FormSectionVersionSummaryDtoArrayApiStatusResult": {
        "type": "object",
        "properties": {
          "successful": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FormSectionVersionSummaryDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublishFormAuthoringDraftRequestDto": {
        "required": [
          "expectedRevision"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "RenameFormSectionRequestDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SaveFormAuthoringDraftRequestDto": {
        "required": [
          "expectedRevision",
          "formData"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "formData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SaveFormSectionAuthoringDraftRequestDto": {
        "required": [
          "expectedRevision",
          "sectionData"
        ],
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "format": "int64"
          },
          "sectionData": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}